Shift comma right; improve G
When a comma followed a 2-pixel tall letter, the comma didn't stand out
as punctuation.

The glyph for G is still awful, but I'm using it because it looks kind
of like a "filled-in" C. And what's the closest thing to a filled-in C?
A capital G!  The next closest thing is probably a lowercase e.

And the glyph kind of does look like a lowercase e...

Dammit. A tic-tac-toe board isn't much to work with.